put pipeline in its own file
